🎉 kickstart project with `instaproxy` Python module
This is a small webserver that acts as a proxy (with cache) for the Instagram GraphQL APIs.
It is powered by [FastAPI](https://github.com/fastapi/fastapi), it is completely asynchronous and does not require many dependencies (not even a database).

A minimal Dockerfile has been included too, the API server will be listening on port 15000 (**not** exposed, as this will be containerised with Docker Compose)